requirements.txt
# dependencies
beautifulsoup4~=4.13.4
colorama~=0.4.6
md2pdf
python-dotenv~=1.1.0
pyyaml
uvicorn~=0.34.2
pydantic~=2.11.4
fastapi~=0.115.12
python-multipart
markdown~=3.8
langchain~=0.3.25
langchain_community
langchain-openai~=0.3.16
tiktoken~=0.9.0
gpt-researcher
arxiv~=2.2.0
PyMuPDF~=1.25.5
requests~=2.32.3
jinja2
aiofiles~=24.1.0
mistune~=3.1.3
python-docx
htmldocx
lxml_html_clean
websockets
unstructured
json_repair~=0.44.1
json5~=0.12.0
loguru~=0.7.3
langgraph~=0.4.2

# uncomment for testing
# pytest
# pytest-asyncio

nest-asyncio~=1.6.0
langchain-community~=0.3.23
tqdm~=4.67.1
langchain-core~=0.3.58
watchdog~=6.0.0
mcp~=1.8.0
click~=8.1.8
typing_extensions~=4.13.2
aiohttp~=3.11.18